MCSC-972 Type changed to Long from Int

+++ app/uk/gov/hmrc/calculatepublicpensionadjustment/config/AppConfig.scala
@@ -28,7 +28,7 @@ class AppConfig @Inject() (config: Configuration, servicesConfig: ServicesConfig
 
   lazy val paacServiceUrl: String = servicesConfig.baseUrl("pension-annual-allowance-calculator")
 
-  val ttlInDays: Int = servicesConfig.getInt("mongodb.ttlInDays")
+  val ttlInDays: Long = servicesConfig.getInt("mongodb.ttlInDays")
 
   val sppaBaseUrl: String =
     config.get[Service]("microservice.services.submit-public-pension-adjustment").baseUrl
